Hai playing piano with the help of Html, CSS, JS is very great full to you.
These files contain the coding to create a piano. Html file contains differ dives of keys in piano. The CSS  file contains decoration of piano. The JS file contains functionalities applied to the keys to play  on mouse and keyboard. The audio function is designed to inject it with keys.